# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: true
    type: object
info:
  contact: {}
  license: {}
paths:
  /balance:
    get:
      description: 返回指定用户的余额信息
      parameters:
      - description: Ivan
        in: query
        name: address
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      summary: 余额
      tags:
      - 前端
  /balancedetailed:
    get:
      description: 返回指定地址的交易明细
      parameters:
      - description: Ivan
        in: query
        name: address
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      summary: 余额明细
      tags:
      - 前端
  /blockchain:
    post:
      description: 用于同步本地区块链数据
      responses:
        "200":
          description: '{"statuc":"ok", "databyte":"bytesdata"}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: 当前区块链数据
      tags:
      - 服务端
  /entry:
    post:
      description: 生产茶叶数据， 用于交易
      parameters:
      - description: Ivan
        in: formData
        name: address
        required: true
        type: string
      - description: "5000"
        in: formData
        name: amount
        required: true
        type: integer
      - description: json
        in: formData
        name: data
        required: true
        type: string
      responses:
        "200":
          description: statuc":"ok", "data":""}
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: statuc":"error", "message":"失败原因"}
          schema:
            $ref: '#/definitions/gin.H'
      summary: 数据录入
      tags:
      - 前端
  /register:
    get:
      description: 注册当前服务器信息到中心服务器中
      parameters:
      - description: "5000"
        in: formData
        name: mode
        required: true
        type: integer
      responses:
        "200":
          description: '{"statuc":"ok"}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: 注册服务器到中心服务器中
      tags:
      - 服务端
  /registerinfo:
    get:
      description: 返回指当前所有的注册信息
      responses:
        "200":
          description: '{"statuc":"ok", "addres": []}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: 返回注册信息
      tags:
      - 服务端
  /teadata:
    post:
      description: 获取指定地址的茶叶数据
      parameters:
      - description: ASHASDSABDKJQWFKJBASFKAF
        in: formData
        name: address
        required: true
        type: string
      responses:
        "200":
          description: statuc":"ok", "data":""}
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: statuc":"error", "message":"失败原因"}
          schema:
            $ref: '#/definitions/gin.H'
      summary: 茶叶数据
      tags:
      - 前端
  /transaction:
    post:
      description: 用于两个不同地址之间的数据交易
      parameters:
      - description: Ivan
        in: formData
        name: from
        required: true
        type: string
      - description: Ble
        in: formData
        name: to
        required: true
        type: string
      - description: "300"
        in: formData
        name: amount
        required: true
        type: integer
      - description: '{}'
        in: formData
        name: data
        type: string
      responses:
        "200":
          description: statuc":"ok"}
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: statuc":"error", "data":"失败原因"}
          schema:
            $ref: '#/definitions/gin.H'
      summary: 茶叶交易
      tags:
      - 前端
  /transactionlist:
    get:
      description: 返回当前区块链里的所有交易数据
      responses:
        "200":
          description: '{"data":{}}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: 区块链数据
      tags:
      - Demo
  /users:
    get:
      description: 获取所有区块链中的用户地址(在实际的区块链中该地址是保密的, 当前为demo演示接口)
      responses:
        "200":
          description: '{"data":["sadhaj","Pedro","Ivan"],"statuc":"ok"}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: 用户数据
      tags:
      - Demo接口
  /version:
    get:
      description: 返回当前区块链长度
      responses:
        "200":
          description: '{"version":20}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: 区块链版本
      tags:
      - 服务端
swagger: "2.0"